requirements.txt
uvicorn
fastapi
python-multipart
langchain==0.0.216
openai==1.2.0
google-cloud-vision